+++ Sources/KeyboardKit/System/SystemKeyboardButtonRowItem.swift
@@ -69,7 +69,10 @@ public struct SystemKeyboardButtonRowItem<Content: View>: View {
             .frame(maxWidth: .infinity)
             .frame(height: item.size.height - item.insets.top - item.insets.bottom)
             .rowItemWidth(for: item, totalWidth: keyboardWidth, referenceWidth: inputWidth)
-            .systemKeyboardButtonStyle(buttonStyle)
+            .systemKeyboardButtonStyle(
+                buttonStyle,
+                isPressed: isPressed
+            )
             .padding(item.insets)
             .background(Color.clearInteractable)
             .keyboardGestures(