+++ RELEASE_NOTES.md
@@ -25,7 +25,9 @@ Due to a bug found in the new autocomplete provider, the old `StandardAutocomple
 * `KeyboardAppearance` has a new `keyboardBackground` property.
 * `KeyboardBackgroundStyle` is a new style that can be used to the background of a keyboard.
 * `KeyboardButtonStyle` properties are now optional to allow for overrides.
+* `KeyboardButtonStyle` has a new `pressedOverlayColor` property that can be used to highlight buttons.
 * `KeyboardButtonStyle` has a new `extended(with:)` function to let you extend a style with another style.
+* `View+SystemKeyboardButton` has a new `isPressed` parameter to let you define pressed state.
 
 ### 👑 Pro changes